First draft of a monitor function
When digging a name, add it to a table which basically looks like rule <-> name <-> IP.
Then add a function that will iterate through this table, resolve the
name to IP addresses, and check if there has been any change. If change
is detected, it will regenerate the associated rules for this name,
using the rule key.